# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API del sistema hospitalario Galenos Pro. Arquitectura hexagonal, persiste en SQL Server 2022.",
        "title": "Galenos Pro Appointments API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/evoluciones/{idRegAtencion}/sintomas": {
            "post": {
                "description": "Reemplaza los síntomas seleccionados de la evolución (sp_go_InsertarEvolucionSintomas)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Evoluciones"
                ],
                "summary": "Guardar síntomas seleccionados de la evolución",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the Registration / Encounter",
                        "name": "idRegAtencion",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Síntomas seleccionados",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpadapter.GuardarSintomasRequest"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/sintomas/catalogo": {
            "get": {
                "description": "Devuelve el catálogo de síntomas agrupado por sistema (Tab_Sintomas_Catalogo)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Evoluciones"
                ],
                "summary": "Listar catálogo de síntomas",
                "responses": {},
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "description": "Inserta un síntoma nuevo en el catálogo si no existe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Evoluciones"
                ],
                "summary": "Agregar síntoma al catálogo",
                "parameters": [
                    {
                        "description": "Síntoma a agregar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpadapter.AgregarSintomaRequest"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "domain.SintomaSeleccionado": {
            "type": "object",
            "properties": {
                "idSintoma": {
                    "type": "integer"
                },
                "sintoma": {
                    "type": "string"
                },
                "sistema": {
                    "type": "string"
                }
            }
        },
...
        "httpadapter.AgregarSintomaRequest": {
            "type": "object",
            "required": [
                "sintoma",
                "sistema"
            ],
            "properties": {
                "sintoma": {
                    "type": "string"
                },
                "sistema": {
                    "type": "string"
                }
            }
        },
...
        "httpadapter.GuardarSintomasRequest": {
            "type": "object",
            "required": [
                "sintomas"
            ],
            "properties": {
                "sintomas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SintomaSeleccionado"
                    }
                }
            }
        },
...
    }
}